feature: change page layout
