+++ src/app/[locale]/(logged-in)/researcher/profile/[tabId]/components/Completion/index.ts
@@ -1,0 +1,3 @@
+import Completion from "./Completion";
+
+export default Completion;
