--- src/app/[locale]/(logged-in)/researcher/profile/[tabId]/components/Details/index.ts
@@ -1,3 +1,0 @@
-import Details from "./Details";
-
-export default Details;
