+++ src/app/[locale]/(logged-in)/researcher/profile/[tabId]/components/Identity/index.ts
@@ -1,3 +1,3 @@
-import Indentity from "./Identity";
+import Identity from "./Identity";
 
-export default Indentity;
+export default Identity;
