--- src/app/[locale]/(logged-in)/researcher/profile/[tabId]/components/Sections/Sections.test.tsx
@@ -1,51 +1,0 @@
-import { useStore } from "@/data/store";
-import { mockedUser } from "@/mocks/data/user";
-import { act, render, screen } from "@/utils/testUtils";
-import { axe } from "jest-axe";
-import Sections from "./Sections";
-
-jest.mock("@/data/store");
-
-const defaultUser = mockedUser();
-
-(useStore as unknown as jest.Mock).mockReturnValue(() => defaultUser);
-
-const renderSections = () => {
-  return render(<Sections>Child</Sections>);
-};
-
-describe("<Sections />", () => {
-  it("has no accessibility validations", async () => {
-    const { container } = renderSections();
-
-    let results;
-
-    await act(async () => {
-      results = await axe(container);
-    });
-
-    expect(results).toHaveNoViolations();
-  });
-
-  it("shows profile is complete", () => {
-    (useStore as unknown as jest.Mock).mockReturnValue(() => ({
-      ...defaultUser,
-      profile_completed_at: "123",
-    }));
-
-    renderSections();
-
-    expect(screen.getByText(/Profile complete/)).toBeInTheDocument();
-  });
-
-  it("shows profile is not complete", () => {
-    (useStore as unknown as jest.Mock).mockReturnValue(() => ({
-      ...defaultUser,
-      profile_completed_at: "",
-    }));
-
-    renderSections();
-
-    expect(screen.getByText(/Profile not complete/)).toBeInTheDocument();
-  });
-});
