--- src/app/[locale]/(logged-in)/researcher/profile/[tabId]/components/Sections/index.ts
@@ -1,3 +1,0 @@
-import Sections from "./Sections";
-
-export default Sections;
