+++ src/app/[locale]/(logged-in)/researcher/profile/[tabId]/consts/tabs.ts
@@ -1,9 +1,9 @@
 enum PageTabs {
-  DETAILS = "details",
   EXPERIENCE = "experience",
   AFFILIATIONS = "affiliations",
   IDENTITY = "identity",
   TRAINING = "training",
+  COMPLETION = "completion",
 }
 
 export { PageTabs };
