+++ src/hooks/useUserProfileCompletion/useUserProfileCompletion.tsx
@@ -2,6 +2,11 @@ import { UserProfileCompletionCategories } from "@/consts/user";
 import { User, UserProfileCompletionSchema } from "@/types/application";
 
 import { useStore } from "@/data/store";
+import { PatchUserPayload } from "@/services/users";
+import { useMutation } from "@tanstack/react-query";
+import patchUser from "@/services/users/patchUser";
+import dayjs from "dayjs";
+import { useCallback } from "react";
 
 const schema: UserProfileCompletionSchema = {
   [UserProfileCompletionCategories.IDENTITY]: {
@@ -16,10 +21,6 @@ const schema: UserProfileCompletionSchema = {
       },
       {
         name: "dob",
-        required: true,
-      },
-      {
-        name: "email",
         required: true,
       },
     ],
@@ -38,35 +39,102 @@ const schema: UserProfileCompletionSchema = {
 export default function useProfileCompletion() {
   const [user, setUser] = useStore(store => [store.getUser(), store.setUser]);
 
-  function update<T>(
-    fields: T,
-    category: UserProfileCompletionCategories,
-    userData?: User
-  ) {
-    let currentState = user?.profile_steps_completed;
-
-    schema[category].fields.forEach(({ name, required }) => {
-      currentState = {
-        ...currentState,
-        [category]: {
-          ...currentState[category],
-          [name]:
-            required &&
-            !!(fields as Record<string, string | boolean | number>)[name],
+  const { mutate, isError, isPending, error } = useMutation({
+    mutationKey: ["patchUser"],
+    mutationFn: (payload: PatchUserPayload) =>
+      patchUser(user?.id, payload, {
+        error: {
+          message: "submitError",
         },
-      };
-    });
+      }),
+  });
 
-    if (user) {
-      setUser({
-        ...user,
-        ...userData,
-        profile_steps_completed: currentState,
-      });
-    }
-  }
+  const isCategoryCompleted = (category: UserProfileCompletionCategories) => {
+    let currentState = JSON.parse(user?.profile_steps_completed || "{}");
+
+    return !Object.keys(currentState[category] || {}).some((key: string) => {
+      return !currentState[category][key];
+    });
+  };
+
+  const update = useCallback(
+    <T,>(
+      formFields: T,
+      category: UserProfileCompletionCategories,
+      userData?: User
+    ) => {
+      let currentState = JSON.parse(user?.profile_steps_completed || "{}");
+      let isCompleted = true;
+
+      Object.keys(schema).forEach(
+        (currentCategory: UserProfileCompletionCategories) => {
+          const { fields } = schema[currentCategory];
+          let score = 0;
+
+          fields.forEach(({ name, required }, i) => {
+            if (currentCategory === category) {
+              const valueCompleted = !!(
+                formFields as Record<string, string | boolean | number>
+              )[name];
+
+              if (required && valueCompleted) {
+                score = Math.round(((i + 1) / fields.length) * 100);
+              }
+
+              currentState = {
+                ...currentState,
+                [category]: {
+                  ...currentState[category],
+                  score,
+                  [name]: required && valueCompleted,
+                },
+              };
+
+              if (!valueCompleted) isCompleted = false;
+            } else {
+              const valueCompleted =
+                currentState[currentCategory][name] || !required;
+
+              currentState = {
+                ...currentState,
+                [currentCategory]: {
+                  ...currentState[currentCategory],
+                  [name]: valueCompleted,
+                },
+              };
+
+              if (!valueCompleted) isCompleted = false;
+            }
+          });
+        }
+      );
+
+      if (user) {
+        const updatedUser = {
+          ...user,
+          ...userData,
+          profile_steps_completed: JSON.stringify(currentState),
+          profile_completed_at: isCompleted ? dayjs().toISOString() : null,
+        };
+
+        mutate(updatedUser);
+        setUser(updatedUser);
+      }
+    },
+    [user]
+  );
+
+  const getJSON = (): UserProfileCompletionSchema => {
+    return JSON.parse(user?.profile_steps_completed || "{}");
+  };
 
   return {
     update,
+    getJSON,
+    isCategoryCompleted,
+    isCompleted: !!user?.profile_completed_at,
+    isError,
+    isLoading: isPending,
+    error,
   };
 }
