+++ src/modules/PageColumnLayout/PageColumnLayout.tsx
@@ -1,26 +1,17 @@
-"use client";
-
-import { Box, useTheme } from "@mui/material";
+import { Grid, GridProps } from "@mui/material";
 import { ReactNode } from "react";
 
-interface PageColumnLayoutProps {
+interface PageColumnLayoutProps extends GridProps {
   children: ReactNode;
 }
 
-export default function PageColumnLayout({ children }: PageColumnLayoutProps) {
-  const theme = useTheme();
-
+export default function PageColumnLayout({
+  children,
+  ...restProps
+}: PageColumnLayoutProps) {
   return (
-    <Box
-      sx={{
-        display: "flex",
-        gap: 5,
-        [theme.breakpoints.down("md")]: {
-          flexDirection: "column",
-          gap: 2,
-        },
-      }}>
+    <Grid container {...restProps}>
       {children}
-    </Box>
+    </Grid>
   );
 }
