+++ src/modules/PageColumnLayout/PageColumnLayoutLeft.tsx
@@ -1,31 +1,17 @@
-"use client";
-
-import { Box, BoxProps, useTheme } from "@mui/material";
+import { GridProps, Grid } from "@mui/material";
 import { ReactNode } from "react";
 
-interface PageColumnLayoutLeftProps extends BoxProps {
+interface PageColumnLayoutLeftProps extends GridProps {
   children: ReactNode;
 }
 
 export default function PageColumnLayoutLeft({
   children,
-  sx,
+  ...restProps
 }: PageColumnLayoutLeftProps) {
-  const theme = useTheme();
-
   return (
-    <Box
-      sx={{
-        display: "flex",
-        alignItems: "center",
-        flexDirection: "column",
-        width: "265px",
-        [theme.breakpoints.down("md")]: {
-          width: "100%",
-        },
-        ...sx,
-      }}>
+    <Grid item md={9} {...restProps}>
       {children}
-    </Box>
+    </Grid>
   );
 }
