+++ src/modules/PageSection/PageSection.styles.ts
@@ -4,6 +4,6 @@ export const StyledPageSection = styled(Box)(
   ({ theme }) => css`
     position: relative;
     z-index: 1;
-    padding: ${theme.spacing(2)} ${theme.spacing(6)};
+    padding: ${theme.spacing(2)};
   `
 );
