+++ src/modules/index.ts
@@ -3,5 +3,6 @@ import Header from "./Header";
 import PageContainer from "./PageContainer";
 import PageTitle from "./PageTitle";
 import PageSection from "./PageSection";
+import PageContent from "./PageContent";
 
-export { Footer, PageContainer, Header, PageTitle, PageSection };
+export { Footer, PageContainer, Header, PageTitle, PageSection, PageContent };
